Extract daysAfter to timify.js
