+++ www/src/js/utils/timify.js
@@ -1,5 +1,5 @@
 // @flow
-import type { DayText, LessonTime, Lesson } from 'types/modules';
+import type { DayText, Lesson, LessonTime } from 'types/modules';
 
 // Converts a 24-hour format time string to an index.
 // Each index corresponds to one cell of each timetable row.
@@ -75,4 +75,13 @@ export function getCurrentDayIndex(
   now: Date = new Date(), // Used for tests only
 ): number {
   return now.getDay() - 1; // Minus 1 because JS week starts on Sunday
+}
+
+/**
+ * Return a copy of the original Date incremented by the given number of days
+ */
+export function daysAfter(startDate: Date, days: number): Date {
+  const d = new Date(startDate.valueOf());
+  d.setUTCDate(d.getUTCDate() + days);
+  return d;
 }
